services/sandbox/pyproject.toml
[project]
name = "forge-sandbox"
version = "12.0.0"
description = "OpenSandbox — Kernel-Level Isolation with Landlock, OPA, SSRF Protection"
requires-python = ">= 3.12"

dependencies = [
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.uv.sources]
forge-security = { path = "../../libs/forge_security", editable = true }
forge-telemetry = { path = "../../libs/forge_telemetry", editable = true }
